Revert away from spurious version increment
