--- packages/notebook-types/examples/v2/SEIRV.json
@@ -1,1 +1,0 @@
-{"type":"model","name":"SEIRV","theory":"primitive-stock-flow","notebook":{"cellContents":{"0194d7a9-d72e-700a-9c8b-7686fad4193b":{"tag":"formal","id":"0194d7a9-d72e-700a-9c8b-7686fad4193b","content":{"tag":"morphism","name":"exposure","id":"0194d7a9-d72e-700a-9c8b-703b753da038","morType":{"tag":"Hom","content":{"tag":"Basic","content":"Object"}},"dom":{"tag":"Basic","content":"0194d7a9-bc26-73ac-b2b8-6eb9a514827b"},"cod":{"tag":"Basic","content":"0194fc13-3da9-77cb-b36a-9b8a6a83f0e6"}}},"0194d7a9-bc26-73ac-b2b8-73470cb412ae":{"tag":"formal","id":"0194d7a9-bc26-73ac-b2b8-73470cb412ae","content":{"tag":"object","name":"Susceptible","id":"0194d7a9-bc26-73ac-b2b8-6eb9a514827b","obType":{"tag":"Basic","content":"Object"}}},"0194d7a9-c036-745b-99e2-b3ef2d3b87d0":{"tag":"formal","id":"0194d7a9-c036-745b-99e2-b3ef2d3b87d0","content":{"tag":"object","name":"Recovered","id":"0194d7a9-c036-745b-99e2-aedb526915c7","obType":{"tag":"Basic","content":"Object"}}},"0194d7a9-bf28-709f-b34d-545928dd7c16":{"tag":"formal","id":"0194d7a9-bf28-709f-b34d-545928dd7c16","content":{"tag":"object","name":"Infectious","id":"0194d7a9-bf28-709f-b34d-513c3da75e9c","obType":{"tag":"Basic","content":"Object"}}},"0194fc14-204c-72a4-92fa-df5491673bab":{"tag":"formal","id":"0194fc14-204c-72a4-92fa-df5491673bab","content":{"tag":"morphism","name":"vaccination","id":"0194fc14-204c-72a4-92fa-da1399afa430","morType":{"tag":"Hom","content":{"tag":"Basic","content":"Object"}},"dom":{"tag":"Basic","content":"0194d7a9-bc26-73ac-b2b8-6eb9a514827b"},"cod":{"tag":"Basic","content":"0194fc13-2920-724c-a6c3-f85ca780f6cf"}}},"0194fc13-d6aa-741c-a30c-54cd59aaee10":{"tag":"formal","id":"0194fc13-d6aa-741c-a30c-54cd59aaee10","content":{"tag":"morphism","name":"infection","id":"0194fc13-d6aa-741c-a30c-536d48c10c1e","morType":{"tag":"Hom","content":{"tag":"Basic","content":"Object"}},"dom":{"tag":"Basic","content":"0194fc13-3da9-77cb-b36a-9b8a6a83f0e6"},"cod":{"tag":"Basic","content":"0194d7a9-bf28-709f-b34d-513c3da75e9c"}}},"0194d7a9-f4ff-77df-80a5-75410f4b9680":{"tag":"formal","id":"0194d7a9-f4ff-77df-80a5-75410f4b9680","content":{"tag":"morphism","name":"","id":"0194d7a9-f4ff-77df-80a5-70d50473ed6d","morType":{"tag":"Basic","content":"Link"},"dom":{"tag":"Basic","content":"0194d7a9-bf28-709f-b34d-513c3da75e9c"},"cod":{"tag":"Tabulated","content":{"tag":"Basic","content":"0194d7a9-d72e-700a-9c8b-703b753da038"}}}},"0194fc13-2920-724c-a6c3-fc1d17699223":{"tag":"formal","id":"0194fc13-2920-724c-a6c3-fc1d17699223","content":{"tag":"object","name":"Vaccinated","id":"0194fc13-2920-724c-a6c3-f85ca780f6cf","obType":{"tag":"Basic","content":"Object"}}},"0194d7a9-d8b5-741f-916f-d559d3ff2f8d":{"tag":"formal","id":"0194d7a9-d8b5-741f-916f-d559d3ff2f8d","content":{"tag":"morphism","name":"recovery","id":"0194d7a9-d8b5-741f-916f-d3dd802450ec","morType":{"tag":"Hom","content":{"tag":"Basic","content":"Object"}},"dom":{"tag":"Basic","content":"0194d7a9-bf28-709f-b34d-513c3da75e9c"},"cod":{"tag":"Basic","content":"0194d7a9-c036-745b-99e2-aedb526915c7"}}},"0194d7c1-3b04-77cf-b2c1-e4e8a10712e2":{"tag":"rich-text","id":"0194d7c1-3b04-77cf-b2c1-e4e8a10712e2","content":"This stock-flow diagram exhibits the standard SIR model from epidemiology augmented with Exposed and Vaccinated populations: there are populations of susceptible, exposed, infectious, vaccinated, and recovered subjects, and processes flow among these stocks. A link from the infectious population to the infection process reflects that the infection rate increases as the density of infectious subjects does."},"0194fc13-3da9-77cb-b36a-9c5b8a74d941":{"tag":"formal","id":"0194fc13-3da9-77cb-b36a-9c5b8a74d941","content":{"tag":"object","name":"Exposed","id":"0194fc13-3da9-77cb-b36a-9b8a6a83f0e6","obType":{"tag":"Basic","content":"Object"}}}},"cellOrder":["0194d7c1-3b04-77cf-b2c1-e4e8a10712e2","0194d7a9-bc26-73ac-b2b8-73470cb412ae","0194fc13-3da9-77cb-b36a-9c5b8a74d941","0194d7a9-bf28-709f-b34d-545928dd7c16","0194d7a9-c036-745b-99e2-b3ef2d3b87d0","0194fc13-2920-724c-a6c3-fc1d17699223","0194d7a9-d72e-700a-9c8b-7686fad4193b","0194fc14-204c-72a4-92fa-df5491673bab","0194fc13-d6aa-741c-a30c-54cd59aaee10","0194d7a9-d8b5-741f-916f-d559d3ff2f8d","0194d7a9-f4ff-77df-80a5-75410f4b9680"]},"version":"2"}
