+++ packages/notebook-types/src/lib.rs
@@ -4,8 +4,7 @@ use serde_wasm_bindgen::{Serializer, from_value};
 use wasm_bindgen::prelude::*;
 
 mod v0;
-mod v1;
-mod v2;
+pub mod v1;
 
 #[cfg(test)]
 mod test_utils;
@@ -13,7 +12,7 @@ mod test_utils;
 pub mod current {
     // this should always track the latest version, and is the only version
     // that is exported from notebook-types
-    pub use crate::v2::*;
+    pub use crate::v1::*;
 }
 
 /// Generate type defs for dependencies supporting `serde` but not `tsify`.
@@ -31,7 +30,7 @@ type Ustr = string;
 type Value = unknown;
 "#;
 
-pub static CURRENT_VERSION: &str = "2";
+pub static CURRENT_VERSION: &str = "1";
 
 #[wasm_bindgen(js_name = "currentVersion")]
 pub fn current_version() -> String {
@@ -42,7 +41,6 @@ pub fn current_version() -> String {
 pub enum VersionedDocument {
     V0(v0::Document),
     V1(v1::Document),
-    V2(v2::Document),
 }
 
 impl<'de> Deserialize<'de> for VersionedDocument {
@@ -65,11 +63,6 @@ impl<'de> Deserialize<'de> for VersionedDocument {
                     serde_json::from_value(value).map_err(serde::de::Error::custom)?;
                 Ok(VersionedDocument::V1(doc))
             }
-            "2" => {
-                let doc: v2::Document =
-                    serde_json::from_value(value).map_err(serde::de::Error::custom)?;
-                Ok(VersionedDocument::V2(doc))
-            }
             other => Err(serde::de::Error::custom(format!("unsupported version {other}"))),
         }
     }
@@ -83,11 +76,7 @@ impl VersionedDocument {
                 VersionedDocument::V1(v1::Document::migrate_from_v0(v0)).to_current()
             }
 
-            VersionedDocument::V1(old1) => {
-                // Recursive call to VersionedNotebook::to_current
-                VersionedDocument::V2(v2::Document::migrate_from_v1(old1)).to_current()
-            }
-            VersionedDocument::V2(old2) => old2,
+            VersionedDocument::V1(old1) => old1,
         }
     }
 }
@@ -118,14 +107,6 @@ mod migration_tests {
     #[test]
     fn test_v0_examples_migrate_to_current() {
         test_example_documents::<VersionedDocument, _>("examples/v0", |doc, _| {
-            // ensure it migrates without panic
-            let _ = doc.to_current();
-        });
-    }
-
-    #[test]
-    fn test_v1_examples_migrate_to_current() {
-        test_example_documents::<VersionedDocument, _>("examples/v1", |doc, _| {
             // ensure it migrates without panic
             let _ = doc.to_current();
         });
