+++ packages/notebook-types/src/v1/notebook.rs
@@ -16,12 +16,12 @@ pub struct Notebook<T> {
     pub cell_order: Vec<Uuid>,
 }
 
-#[derive(PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[allow(dead_code)]
+#[derive(PartialEq, Eq, Debug, Serialize, Deserialize, Tsify)]
+#[tsify(into_wasm_abi, from_wasm_abi)]
 pub struct ModelNotebook(pub Notebook<super::model_judgment::ModelJudgment>);
 
-#[derive(PartialEq, Eq, Debug, Serialize, Deserialize)]
-#[allow(dead_code)] //should probably delete
+#[derive(PartialEq, Eq, Debug, Serialize, Deserialize, Tsify)]
+#[tsify(into_wasm_abi, from_wasm_abi)]
 pub struct DiagramNotebook(pub Notebook<super::diagram_judgment::DiagramJudgment>);
 
 impl<T> Notebook<T> {
